# Conflicts:
#	insty-api/src/test/java/insty/domain/community/service/CommunityServiceTest.java

+++ insty-api/src/main/java/insty/domain/course/service/CourseService.java
@@ -9,6 +9,8 @@ import insty.domain.course.dto.CourseDetailRes;
 import insty.domain.course.dto.CourseInstallEnvChecklistInfo;
 import insty.domain.course.dto.CourseMySearchInfo;
 import insty.domain.course.dto.CourseMySearchReq;
+import insty.domain.course.dto.CourseRequestReq;
+import insty.domain.course.dto.CourseRequestRes;
 import insty.domain.course.dto.CourseSearchFilter;
 import insty.domain.course.dto.CourseSearchInfo;
 import insty.domain.course.dto.CourseSearchReq;
@@ -18,12 +20,18 @@ import insty.domain.course.implement.CourseCounter;
 import insty.domain.course.implement.CourseFileReader;
 import insty.domain.course.implement.CourseFileWriter;
 import insty.domain.course.implement.CourseReader;
+import insty.domain.course.implement.CourseRequestReader;
+import insty.domain.course.implement.CourseRequestWriter;
 import insty.domain.course.implement.CourseTagWriter;
 import insty.domain.course.implement.CourseValidator;
 import insty.domain.course.implement.CourseVideoManager;
 import insty.domain.course.implement.CourseWriter;
+import insty.domain.user.implement.UserReader;
 import insty.model.course.Course;
+import insty.model.course.CourseRequest;
+import insty.model.user.User;
 import java.util.List;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -40,29 +48,37 @@ public class CourseService {
     private final CourseFileWriter courseFileWriter;
     private final CourseFileReader courseFileReader;
     private final CourseTagWriter courseTagWriter;
+    private final CourseRequestWriter courseRequestWriter;
+    private final CourseRequestReader courseRequestReader;
     private final CourseVideoManager courseVideoManager;
     private final CourseValidator courseValidator;
     private final CourseComplexReader courseComplexReader;
+    private final UserReader userReader;
 
-    public CourseDetailRes createCourse(CourseCreateReq req, MultipartFile thumbnail,
+    public CourseDetailRes createCourse(Long userId, CourseCreateReq req, MultipartFile thumbnail,
                                         List<MultipartFile> practiceFile) {
-        Course course = courseWriter.saveCourse(req);
-        courseVideoManager.attachmentCourse(course, req.videoUuid());
-        String thumbnailUrl = courseFileWriter.saveThumbnailAndGetUrl(thumbnail, course);
+        User user = userReader.getUser(userId);
+        Course course = courseWriter.saveCourse(user, req);
+        UUID videoUuid = courseVideoManager.attachmentCourse(course, req.videoUuid());
+        courseFileWriter.saveThumbnail(thumbnail, course);
+        String thumbnailUrl = courseFileReader.getThumbnailUrl(course, videoUuid);
         List<FileInfo> practiceFileInfos = courseFileWriter.savePracticeFilesAndGetInfo(practiceFile, course);
         List<CourseInstallEnvChecklistInfo> checklists = courseWriter.saveCourseInstallEnvChecklist(course,
                 req.installEnvChecklist());
         List<String> keypoints = courseWriter.saveCourseKeypoints(course, req.keyPoints());
         List<String> tags = courseTagWriter.saveCourseTagsAndGetTagNames(course, req.tags());
 
-        return CourseDetailRes.from(course, checklists, keypoints, tags, thumbnailUrl, practiceFileInfos);
+        return CourseDetailRes.from(course, checklists, keypoints, tags, thumbnailUrl, practiceFileInfos,
+                videoUuid);
     }
 
-    public CourseDetailRes updateCourse(Long courseId, CourseUpdateReq req, MultipartFile thumbnail,
+    public CourseDetailRes updateCourse(Long userId, Long courseId, CourseUpdateReq req, MultipartFile thumbnail,
                                         List<MultipartFile> practiceFile) {
+        courseValidator.validateCourseOwner(courseId, userId);
         Course course = courseWriter.updateCourse(courseId, req);
-        courseVideoManager.updateVideo(course, req.updateVideoUuid());
-        String thumbnailUrl = courseFileWriter.updateThumbnailAndGetUrl(thumbnail, course);
+        UUID videoUuid = courseVideoManager.updateVideo(course, req.updateVideoUuid());
+        courseFileWriter.updateThumbnail(thumbnail, course);
+        String thumbnailUrl = courseFileReader.getThumbnailUrl(course, videoUuid);
         List<FileInfo> fileInfos = courseFileWriter.updatePracticeFilesAndGetInfo(practiceFile,
                 req.deletePracticeFileId(), course);
         List<CourseInstallEnvChecklistInfo> checklists = courseWriter.updateCourseInstallEnvChecklist(course,
@@ -70,7 +86,7 @@ public class CourseService {
         List<String> keypoints = courseWriter.updateCourseKeypoints(course, req.keyPoints());
         List<String> tags = courseTagWriter.updateCourseTags(course, req.tags());
 
-        return CourseDetailRes.from(course, checklists, keypoints, tags, thumbnailUrl, fileInfos);
+        return CourseDetailRes.from(course, checklists, keypoints, tags, thumbnailUrl, fileInfos, videoUuid);
     }
 
     /**
@@ -91,10 +107,11 @@ public class CourseService {
         List<CourseInstallEnvChecklistInfo> checklists = courseReader.getChecklistsByCourseId(course.getId());
         List<String> keypoints = courseReader.getKeypointContentsByCourseId(course.getId());
         List<String> tagNames = courseReader.getTagNamesByCourseId(course.getId());
-        String thumbnailUrl = courseFileReader.getThumbnailUrl(course);
+        UUID videoUuid = courseVideoManager.getAttachVideoUuid(course.getId());
+        String thumbnailUrl = courseFileReader.getThumbnailUrl(course, videoUuid);
         List<FileInfo> practiceFiles = courseFileReader.getPracticeFiles(course);
 
-        return CourseDetailRes.from(course, checklists, keypoints, tagNames, thumbnailUrl, practiceFiles);
+        return CourseDetailRes.from(course, checklists, keypoints, tagNames, thumbnailUrl, practiceFiles, videoUuid);
     }
 
     public SearchRes<CourseSearchInfo> searchCourse(CourseSearchReq req) {
@@ -102,6 +119,7 @@ public class CourseService {
         CourseSearchFilter filter = req.toSearchFilter();
 
         List<CourseSearchInfo> searchInfo = courseComplexReader.searchCourse(paginationReq, filter);
+        searchInfo = courseComplexReader.setBasicThumbnailUrlForSearch(searchInfo);
         PaginationRes paginationRes = courseComplexReader.countSearchCourse(paginationReq, filter);
 
         return SearchRes.from(paginationRes, searchInfo);
@@ -111,8 +129,25 @@ public class CourseService {
         PaginationReq paginationReq = req.toPaginationReq();
 
         List<CourseMySearchInfo> searchInfo = courseComplexReader.searchMyCourse(paginationReq, userId);
+        searchInfo = courseComplexReader.setBasicThumbnailUrlForMy(searchInfo);
         PaginationRes paginationRes = courseComplexReader.countSearchMyCourse(paginationReq, userId);
 
         return SearchRes.from(paginationRes, searchInfo);
+    }
+
+    /**
+     * 러너가 크리에이터에게 강의 요청
+     */
+    public CourseRequestRes createCourseRequest(Long userId, CourseRequestReq req) {
+        CourseRequest saveCourseRequest = courseRequestWriter.saveCourseRequest(userId, req);
+        return CourseRequestRes.from(saveCourseRequest);
+    }
+
+    /**
+     * 크리에이터에 요청된 강의 목록 조회
+     */
+    public List<CourseRequestRes> searchCourseRequest(Long userId) {
+        List<CourseRequest> findMyCourseRequest = courseRequestReader.getListMyCourseRequest(userId);
+        return CourseRequestRes.from(findMyCourseRequest);
     }
 }
